requirements.txt
awesome-slugify==1.6.5
beautifulsoup4==4.8.2
bibtexparser==1.1.0
clld==6.0.0
clldmpg==3.5.0
clldutils==3.5.0
numpy==1.17.0
pandas==0.25.0